backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-webpush==0.3.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==25.1.0
http_ece==1.2.1
idna==3.11
multidict==6.7.1
packaging==26.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-vapid==1.9.4
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
pywebpush==2.3.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
yarl==1.23.0
